# backend/openapi.json
{
  "openapi": "3.1.0",
  "info": {
    "title": "Open-Q API",
    "version": "0.1.0"
  },
  "paths": {
...
    "/api/admin/workspaces": {
      "get": {
        "tags": [
          "admin-workspaces"
        ],
        "summary": "List Workspaces",
        "description": "List all workspaces the current user is a member of, with their role.",
        "operationId": "list_workspaces_api_admin_workspaces_get",
        "responses": {
          "200": {
            "description": "Successful Response",
            "content": {
              "application/json": {
                "schema": {
                  "items": {
                    "$ref": "#/components/schemas/WorkspaceWithRole"
                  },
                  "type": "array",
                  "title": "Response List Workspaces Api Admin Workspaces Get"
                }
              }
            }
          }
        },
        "security": [
          {
            "OAuth2PasswordBearer": []
          }
        ]
      },
      "post": {
        "tags": [
          "admin-workspaces"
        ],
        "summary": "Create Workspace",
        "description": "Create a new workspace and assign the current user as Admin.",
        "operationId": "create_workspace_api_admin_workspaces_post",
        "requestBody": {
          "content": {
            "application/json": {
              "schema": {
                "$ref": "#/components/schemas/WorkspaceCreate"
              }
            }
          },
          "required": true
        },
        "responses": {
          "201": {
            "description": "Successful Response",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/WorkspaceRead"
                }
              }
            }
          },
          "422": {
            "description": "Validation Error",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/HTTPValidationError"
                }
              }
            }
          }
        },
        "security": [
          {
            "OAuth2PasswordBearer": []
          }
        ]
      }
    },
    "/api/admin/workspaces/{slug}": {
      "get": {
        "tags": [
          "admin-workspaces"
        ],
        "summary": "Get Workspace",
        "description": "Get workspace details.",
        "operationId": "get_workspace_api_admin_workspaces__slug__get",
        "security": [
          {
            "OAuth2PasswordBearer": []
          }
        ],
        "parameters": [
          {
            "name": "slug",
            "in": "path",
            "required": true,
            "schema": {
              "type": "string",
              "description": "The slug of the workspace",
              "title": "Slug"
            },
            "description": "The slug of the workspace"
          }
        ],
        "responses": {
          "200": {
            "description": "Successful Response",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/WorkspaceRead"
                }
              }
            }
          },
          "422": {
            "description": "Validation Error",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/HTTPValidationError"
                }
              }
            }
          }
        }
      },
      "patch": {
        "tags": [
          "admin-workspaces"
        ],
        "summary": "Update Workspace",
        "description": "Update workspace details.",
        "operationId": "update_workspace_api_admin_workspaces__slug__patch",
        "security": [
          {
            "OAuth2PasswordBearer": []
          }
        ],
        "parameters": [
          {
            "name": "slug",
            "in": "path",
            "required": true,
            "schema": {
              "type": "string",
              "description": "The slug of the workspace",
              "title": "Slug"
            },
            "description": "The slug of the workspace"
          }
        ],
        "requestBody": {
          "required": true,
          "content": {
            "application/json": {
              "schema": {
                "$ref": "#/components/schemas/WorkspaceUpdate"
              }
            }
          }
        },
        "responses": {
          "200": {
            "description": "Successful Response",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/WorkspaceRead"
                }
              }
            }
          },
          "422": {
            "description": "Validation Error",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/HTTPValidationError"
                }
              }
            }
          }
        }
      },
      "delete": {
        "tags": [
          "admin-workspaces"
        ],
        "summary": "Delete Workspace",
        "description": "Delete a workspace (Owner only).",
        "operationId": "delete_workspace_api_admin_workspaces__slug__delete",
        "security": [
          {
            "OAuth2PasswordBearer": []
          }
        ],
        "parameters": [
          {
            "name": "slug",
            "in": "path",
            "required": true,
            "schema": {
              "type": "string",
              "description": "The slug of the workspace",
              "title": "Slug"
            },
            "description": "The slug of the workspace"
          }
        ],
        "responses": {
          "204": {
            "description": "Successful Response"
          },
          "422": {
            "description": "Validation Error",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/HTTPValidationError"
                }
              }
            }
          }
        }
      }
    },
    "/api/admin/workspaces/{slug}/members": {
      "get": {
        "tags": [
          "admin-workspaces"
        ],
        "summary": "List Workspace Members",
        "description": "List all members of a workspace.",
        "operationId": "list_workspace_members_api_admin_workspaces__slug__members_get",
        "security": [
          {
            "OAuth2PasswordBearer": []
          }
        ],
        "parameters": [
          {
            "name": "slug",
            "in": "path",
            "required": true,
            "schema": {
              "type": "string",
              "description": "The slug of the workspace",
              "title": "Slug"
            },
            "description": "The slug of the workspace"
          }
        ],
        "responses": {
          "200": {
            "description": "Successful Response",
            "content": {
              "application/json": {
                "schema": {
                  "type": "array",
                  "items": {
                    "$ref": "#/components/schemas/WorkspaceMemberRead"
                  },
                  "title": "Response List Workspace Members Api Admin Workspaces  Slug  Members Get"
                }
              }
            }
          },
          "422": {
            "description": "Validation Error",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/HTTPValidationError"
                }
              }
            }
          }
        }
      }
    },
    "/api/admin/workspaces/{slug}/members/{user_id}": {
      "patch": {
        "tags": [
          "admin-workspaces"
        ],
        "summary": "Update Workspace Member",
        "description": "Update a workspace member's role.",
        "operationId": "update_workspace_member_api_admin_workspaces__slug__members__user_id__patch",
        "security": [
          {
            "OAuth2PasswordBearer": []
          }
        ],
        "parameters": [
          {
            "name": "user_id",
            "in": "path",
            "required": true,
            "schema": {
              "type": "integer",
              "title": "User Id"
            }
          },
          {
            "name": "slug",
            "in": "path",
            "required": true,
            "schema": {
              "type": "string",
              "description": "The slug of the workspace",
              "title": "Slug"
            },
            "description": "The slug of the workspace"
          }
        ],
        "requestBody": {
          "required": true,
          "content": {
            "application/json": {
              "schema": {
                "$ref": "#/components/schemas/WorkspaceMemberUpdate"
              }
            }
          }
        },
        "responses": {
          "200": {
            "description": "Successful Response",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/WorkspaceMemberRead"
                }
              }
            }
          },
          "422": {
            "description": "Validation Error",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/HTTPValidationError"
                }
              }
            }
          }
        }
      },
      "delete": {
        "tags": [
          "admin-workspaces"
        ],
        "summary": "Remove Workspace Member",
        "description": "Remove a member from the workspace.",
        "operationId": "remove_workspace_member_api_admin_workspaces__slug__members__user_id__delete",
        "security": [
          {
            "OAuth2PasswordBearer": []
          }
        ],
        "parameters": [
          {
            "name": "user_id",
            "in": "path",
            "required": true,
            "schema": {
              "type": "integer",
              "title": "User Id"
            }
          },
          {
            "name": "slug",
            "in": "path",
            "required": true,
            "schema": {
              "type": "string",
              "description": "The slug of the workspace",
              "title": "Slug"
            },
            "description": "The slug of the workspace"
          }
        ],
        "responses": {
          "204": {
            "description": "Successful Response"
          },
          "422": {
            "description": "Validation Error",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/HTTPValidationError"
                }
              }
            }
          }
        }
      }
    },
    "/api/admin/workspaces/{slug}/invitations": {
      "post": {
        "tags": [
          "admin-workspaces"
        ],
        "summary": "Create Invitation",
        "description": "Invite a user to the workspace.",
        "operationId": "create_invitation_api_admin_workspaces__slug__invitations_post",
        "security": [
          {
            "OAuth2PasswordBearer": []
          }
        ],
        "parameters": [
          {
            "name": "slug",
            "in": "path",
            "required": true,
            "schema": {
              "type": "string",
              "description": "The slug of the workspace",
              "title": "Slug"
            },
            "description": "The slug of the workspace"
          }
        ],
        "requestBody": {
          "required": true,
          "content": {
            "application/json": {
              "schema": {
                "$ref": "#/components/schemas/WorkspaceInvitationCreate"
              }
            }
          }
        },
        "responses": {
          "200": {
            "description": "Successful Response",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/InvitationLink"
                }
              }
            }
          },
          "422": {
            "description": "Validation Error",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/HTTPValidationError"
                }
              }
            }
          }
        }
      }
    },
...
  },
  "components": {
    "schemas": {
...
      "HTTPValidationError": {
        "properties": {
          "detail": {
            "items": {
              "$ref": "#/components/schemas/ValidationError"
            },
            "type": "array",
            "title": "Detail"
          }
        },
        "type": "object",
        "title": "HTTPValidationError"
      },
...
      "InvitationLink": {
        "properties": {
          "invite_url": {
            "type": "string",
            "title": "Invite Url"
          },
          "token": {
            "type": "string",
            "title": "Token"
          }
        },
        "type": "object",
        "required": [
          "invite_url",
          "token"
        ],
        "title": "InvitationLink",
        "description": "Schema for returning a generated invitation link."
      },
...
      "UserRead": {
        "properties": {
          "email": {
            "type": "string",
            "title": "Email"
          },
          "full_name": {
            "anyOf": [
              {
                "type": "string"
              },
              {
                "type": "null"
              }
            ],
            "title": "Full Name"
          },
          "id": {
            "type": "integer",
            "title": "Id"
          },
          "is_active": {
            "type": "boolean",
            "title": "Is Active"
          },
          "is_superuser": {
            "type": "boolean",
            "title": "Is Superuser"
          },
          "is_totp_enabled": {
            "type": "boolean",
            "title": "Is Totp Enabled"
          }
        },
        "type": "object",
        "required": [
          "email",
          "id",
          "is_active",
          "is_superuser",
          "is_totp_enabled"
        ],
        "title": "UserRead",
        "description": "Schema for reading user details."
      },
...
      "ValidationError": {
        "properties": {
          "loc": {
            "items": {
              "anyOf": [
                {
                  "type": "string"
                },
                {
                  "type": "integer"
                }
              ]
            },
            "type": "array",
            "title": "Location"
          },
          "msg": {
            "type": "string",
            "title": "Message"
          },
          "type": {
            "type": "string",
            "title": "Error Type"
          }
        },
        "type": "object",
        "required": [
          "loc",
          "msg",
          "type"
        ],
        "title": "ValidationError"
      },
      "WorkspaceCreate": {
        "properties": {
          "title": {
            "type": "string",
            "title": "Title"
          },
          "slug": {
            "type": "string",
            "maxLength": 50,
            "minLength": 3,
            "pattern": "^[a-z0-9-]+$",
            "title": "Slug"
          }
        },
        "type": "object",
        "required": [
          "title",
          "slug"
        ],
        "title": "WorkspaceCreate",
        "description": "Schema for creating a workspace."
      },
      "WorkspaceInvitationCreate": {
        "properties": {
          "email": {
            "type": "string",
            "title": "Email"
          },
          "role": {
            "$ref": "#/components/schemas/WorkspaceRole",
            "default": "researcher"
          }
        },
        "type": "object",
        "required": [
          "email"
        ],
        "title": "WorkspaceInvitationCreate",
        "description": "Schema for creating a workspace invitation."
      },
      "WorkspaceMemberRead": {
        "properties": {
          "user_id": {
            "type": "integer",
            "title": "User Id"
          },
          "user": {
            "$ref": "#/components/schemas/UserRead"
          },
          "role": {
            "$ref": "#/components/schemas/WorkspaceRole"
          },
          "joined_at": {
            "type": "string",
            "format": "date-time",
            "title": "Joined At"
          }
        },
        "type": "object",
        "required": [
          "user_id",
          "user",
          "role",
          "joined_at"
        ],
        "title": "WorkspaceMemberRead",
        "description": "Schema for reading workspace member details."
      },
      "WorkspaceMemberUpdate": {
        "properties": {
          "role": {
            "$ref": "#/components/schemas/WorkspaceRole"
          }
        },
        "type": "object",
        "required": [
          "role"
        ],
        "title": "WorkspaceMemberUpdate",
        "description": "Schema for updating a workspace member."
      },
      "WorkspaceRead": {
        "properties": {
          "id": {
            "type": "integer",
            "title": "Id"
          },
          "title": {
            "type": "string",
            "title": "Title"
          },
          "slug": {
            "type": "string",
            "title": "Slug"
          },
          "created_at": {
            "type": "string",
            "format": "date-time",
            "title": "Created At"
          },
          "members": {
            "items": {
              "$ref": "#/components/schemas/WorkspaceMemberRead"
            },
            "type": "array",
            "title": "Members",
            "default": []
          }
        },
        "type": "object",
        "required": [
          "id",
          "title",
          "slug",
          "created_at"
        ],
        "title": "WorkspaceRead",
        "description": "Schema for reading a workspace."
      },
      "WorkspaceRole": {
        "type": "string",
        "enum": [
          "owner",
          "admin",
          "researcher",
          "viewer"
        ],
        "title": "WorkspaceRole",
        "description": "Enum for workspace roles."
      },
      "WorkspaceUpdate": {
        "properties": {
          "title": {
            "anyOf": [
              {
                "type": "string"
              },
              {
                "type": "null"
              }
            ],
            "title": "Title"
          },
          "slug": {
            "anyOf": [
              {
                "type": "string",
                "maxLength": 50,
                "minLength": 3,
                "pattern": "^[a-z0-9-]+$"
              },
              {
                "type": "null"
              }
            ],
            "title": "Slug"
          }
        },
        "type": "object",
        "title": "WorkspaceUpdate",
        "description": "Schema for updating a workspace."
      },
      "WorkspaceWithRole": {
        "properties": {
          "id": {
            "type": "integer",
            "title": "Id"
          },
          "title": {
            "type": "string",
            "title": "Title"
          },
          "slug": {
            "type": "string",
            "title": "Slug"
          },
          "created_at": {
            "type": "string",
            "format": "date-time",
            "title": "Created At"
          },
          "members": {
            "items": {
              "$ref": "#/components/schemas/WorkspaceMemberRead"
            },
            "type": "array",
            "title": "Members",
            "default": []
          },
          "user_role": {
            "$ref": "#/components/schemas/WorkspaceRole"
          }
        },
        "type": "object",
        "required": [
          "id",
          "title",
          "slug",
          "created_at",
          "user_role"
        ],
        "title": "WorkspaceWithRole",
        "description": "Schema for reading a workspace with the current user's role."
      }
    },
    "securitySchemes": {
      "OAuth2PasswordBearer": {
        "type": "oauth2",
        "flows": {
          "password": {
            "scopes": {},
            "tokenUrl": "api/token"
          }
        }
      }
    }
  }
}